adding hotebeds data import

+++ api/models/models.py
@@ -51,6 +51,7 @@ class bookingrequest(models.Model):
     package = models.BooleanField(default=False)  # <--
     Standalone = models.BooleanField(default=True)
 
+
 class mappingcodes(models.Model):
 
     class Meta:
@@ -59,12 +60,12 @@ class mappingcodes(models.Model):
     provider_id = models.IntegerField(5)
     hotel_codes = models.IntegerField(1, default=1)
     hotel_name = models.CharField(max_length=50)
-    category_name = models.CharField(max_length=20, blank = True)
+    category_name = models.CharField(max_length=20, blank=True)
     chain_name = models.CharField(max_length=50)
     country_name = models.CharField(max_length=50)
     destination_name = models.CharField(max_length=50)
     address = models.CharField(max_length=75)
     postal_code = models.CharField(max_length=10)
     city = models.CharField(max_length=50)
-    latitude = models.DecimalField(decimal_places=6, max_digits=11)
-    longitude = models.DecimalField(decimal_places=6, max_digits=10)
+    latitude = models.FloatField()
+    longitude = models.FloatField()
